requirements.txt
PyYAML==3.12
pymongo==3.7.2
requests==2.21.0
beautifulsoup4==4.7.1
